libcamera2: 1080p fps with DVS on/off cannot hit 30
BZ: 49330

Issue:
With DVS enable the FPS in 1080P recording can't reach 30fps

Solution:
When recording in 1080P, change preview size to 640x360

In R3, we have the same issue caused by ISP performance limitation, the solution
is to reduce the preview size in 1080p recording when DVS enable, it's workaround.
So in R4, we will use the same solution, judge if the current state is in 1080p
recording preview and dvs enable, if yes, we will set the resolution to 640x360
as preview size.

Change-Id: I0d85e81839d6a68b912a8a6df64c9fa70c5f8957
Signed-off-by: hongjian <hongjiang.a.zheng@intel.com>
Reviewed-on: http://android.intel.com:8080/62030
Reviewed-by: Koski, Anttu <anttu.koski@intel.com>
Tested-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>