libcamera2: Move DVS into its own class
BZ: 51281 53190

This patch moves the DVS glue code out of libmfldadvci/src/3a.c into
a class (AtomDvs) in libcamera2.
This is part of the IA imaging library creation where libmfldadvci will
become an OS independent library and all Android specifics will be contained
in libcamera2. This will allow us to reuse this imaging library for other
platforms as well and this will also draw a clear line that tells which
engineering teams own what functionality.

Change-Id: I361dd67e488b62f2a221c49fc75b04bda3f36309
Signed-off-by: Jozef Kruger <jozef.kruger@intel.com>
Reviewed-on: http://android.intel.com:8080/62142
Reviewed-by: Koski, Anttu <anttu.koski@intel.com>
Tested-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: Verdu, Miguel <miguel.verdu@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>