libcamera2: buffer sharing for hw jpeg encoding
BZ: 49630

Issue:
For performance reason, we should not copy our camera data
to libva if we want to do hw jpeg encoding.

Solution:
This patch is used to enable the buffer sharing for hw jpeg encoding.
The camera will pass one buffer to libva, the libva will encode
that buffer's raw data to jpeg.

Depends on:
61945, 61946

Change-Id: I4d2c928ae18627f45b728aae7a0c8c5d15d4965f
Signed-off-by: lyang56 <liang.l.yang@intel.com>
Signed-off-by: Miguel Verdu <miguel.verdu@intel.com>
Reviewed-on: http://android.intel.com:8080/61947
Reviewed-by: Vehmanen, Kai <kai.vehmanen@intel.com>
Reviewed-by: Zheng, ZixiangX <zixiangx.zheng@intel.com>
Tested-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>