libcamera2: do not use flash in burst captures
BZ: 48083

We do not allow flash to be used in burst and continuous shooting modes.
Instead we only offer the torch mode that’s also offered in the video
mode. Just like in video mode, the auto-flash mode will not be
available. The user will have to manually enable the torch to use flash
with burst and continuous shooting.

Change-Id: Ife9adbaabd030cd748263d83e6a3d8761d0242ca
Signed-off-by: Kai Vehmanen <kai.vehmanen@intel.com>
Reviewed-on: http://android.intel.com:8080/62521
Reviewed-by: Koski, Anttu <anttu.koski@intel.com>
Tested-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>