libcamera2: refresh settings for manual/tv ae mode
BZ: 48105

When AE mode was switched to 'manual' or 'shutter-priority', the
currently set shutter value was not passed to ia_3a library. To really
activate the new AE mode, application needed to change the shutter value
to some new value to refresh the state.

Fix the issue by storing cached values of shutter, and passing that to
ia_3a in case manual/shutter-priorty ae mode is selected.

Change-Id: I1049d2d7d9112df19c2cc3c70ad841eb61cfa2e1
Signed-off-by: Kai Vehmanen <kai.vehmanen@intel.com>
Reviewed-on: http://android.intel.com:8080/63438
Reviewed-by: Verdu, Miguel <miguel.verdu@intel.com>
Reviewed-by: Lampila, KalleX <kallex.lampila@intel.com>
Reviewed-by: Niemimuukko, Ukri <ukri.niemimuukko@intel.com>
Reviewed-by: Alanen, Juha <juha.alanen@intel.com>
Tested-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>