libcamera2: Allocate snapshotbuffers in PictureThread.
BZ: 54572

As part of the performance optimizations for JPEG encoding
we allocate now the snapshotbuffers in the PictureThread.
They will be passed to the ControlThread before initializing
the device for capture.
The allocation of the big buffers does not happen anymore during
take_picture but in preview when parameters are set.

The semantics between CtrlThread and PictureThread are also simplified
now there is no need to return buffers back to PictureThread.

Change-Id: I8b103a0a37e2bc1edc2f6e66b2fa98e8fdc985da
Signed-off-by: Miguel Verdu <miguel.verdu@intel.com>
Reviewed-on: http://android.intel.com:8080/63874
Reviewed-by: Vehmanen, Kai <kai.vehmanen@intel.com>
Reviewed-by: Yang, Liang L <liang.l.yang@intel.com>
Reviewed-by: Koski, Anttu <anttu.koski@intel.com>
Tested-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>