libcamera2: Move HW encoder to PictureThread
BZ: 54572

As part of the performance optimizations for JPEG encoding
we move the HW encoder class out from under JpegCompressor
into the PictureThread

We now create the libVA context when the buffers are allocated by the
PictureThread thus cutting down time to setup the VA context for each encode
request.
In case we receive a buffer that is not part of the current context
we create a temporary libVA context of 1 surface.This is used mainly
for HDR case.

Also we avoid copying the SOI and EOI markers created by the HW encoder. We
currently do not use them

Change-Id: I6f7be8fbfcfe19b6868cd5a2b9866dc8cf7590fc
Signed-off-by: Miguel Verdu <miguel.verdu@intel.com>
Reviewed-on: http://android.intel.com:8080/63875
Reviewed-by: Vehmanen, Kai <kai.vehmanen@intel.com>
Reviewed-by: Koski, Anttu <anttu.koski@intel.com>
Tested-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>