libcamera2: implement interleaved encoding of thumb and main images
BZ: 54572

Last part of the HW encoder performance optimizations
We make use of the asynchronous interface of the JPEG HW encoder to interleave
the encoding process of the main picture done by the HW encoder with the
encoding process of the thumbnail done by the SW encoding.

Achieving optimizations on the total time of encoding.

Change-Id: I7198973ebbb4394e970ecb65f06d46b3637448b8
Signed-off-by: Miguel Verdu <miguel.verdu@intel.com>
Reviewed-on: http://android.intel.com:8080/63876
Reviewed-by: Vehmanen, Kai <kai.vehmanen@intel.com>
Reviewed-by: Koski, Anttu <anttu.koski@intel.com>
Tested-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>