libcamera2: use driver ts now that they are monotonic
BZ: 55081

In video recording we were using the timestamps from the driver
for accuracy, but because they came from gettimeofday() they could
change at any time. To try to compensate we had in the hal the variable
  mTimeRealMonoInterval
that was trying to compensate this (unseccessfully)
Now the driver has changed and it uses ktime_get_ts() which provides
monotonically increasing ts.

We do not need anymore this field in AtomISP.

Change-Id: I5a232f8d1902941fd6e8c1ad488482687a89bbb2
Signed-off-by: Miguel Verdu <miguel.verdu@intel.com>
Reviewed-on: http://android.intel.com:8080/64658
Reviewed-by: Lampila, KalleX <kallex.lampila@intel.com>
Reviewed-by: Koski, Anttu <anttu.koski@intel.com>
Tested-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>