libcamera2: activate fps debug with property
BZ: 57092

Preview Frame Rate debugging has been latent feature in the HAL
that we have not used so far.
It needed to be enabled at build time
We have now the property we use to change the log level at
start up.

This patch enables this feature  when the camera.hal.debug property
is set to 128 or has that bit on.

Change-Id: I5b2cfe95f3f9f5e527d3dc19a1db4f4edf329700
Signed-off-by: Miguel Verdu <miguel.verdu@intel.com>
Reviewed-on: http://android.intel.com:8080/65886
Reviewed-by: Niemimuukko, Ukri <ukri.niemimuukko@intel.com>
Reviewed-by: Hu, Gang A <gang.a.hu@intel.com>
Reviewed-by: Lampila, KalleX <kallex.lampila@intel.com>
Reviewed-by: Vehmanen, Kai <kai.vehmanen@intel.com>
Tested-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>