libcamera2: Fixes landscape infinity focus mode
BZ: 49149

The landscape mode did not use a proper focus distance for
infinity focus. With this patch we get the focus value for
infinity from 3A library, and set that value as manual focus
value. Also, taking into use a set of flags for restricting which
functions of 3A is used with face detection in PostProcThread.

Change-Id: Ie433ca857d4cd3ad44c6452310d739853dd5cf65
Signed-off-by: Lassi Väätämöinen <lassix.vtminen@intel.com>
Reviewed-on: http://android.intel.com:8080/65888
Reviewed-by: Hu, Gang A <gang.a.hu@intel.com>
Tested-by: Hu, Gang A <gang.a.hu@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>