libcamera2: enable HDR with Smart Shutter
BZ: 58781

In HDR + smart shutter case, with SaveOriginal selected,
the first takePicture sent by the app needs to be delayed
because of the smart shutter requirement.
The second takePicture sent by the application
needed to retrieve the original image, should be processed
directly.
This patch is taking care of this matter by introducing
a flag, which indicate which takePicture routine, libcamera
should take.

Change-Id: Ifc9f245444a58f6988246d87a17d41fcea50668a
Signed-off-by: Arnaud Mandy <arnaud.mandy@intel.com>
Reviewed-on: http://android.intel.com:8080/67342
Reviewed-by: Vehmanen, Kai <kai.vehmanen@intel.com>
Tested-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>