libcamera2: consider buffer stride when encoding JPEG in SW
BZ: 59098

SW JPEG encoder is used for SW encoding which is used for video
snapshot
For some video resolutions the ISP allocates a bigger buffer than
requested because the writes per line need to be an integer number
of 256 bits (i.e. 32 bytes), the ISP bus width.
So all lines need to be multiple of 32.

When this type of buffer reach the JPEG encoder
the encoder only uses the declared width not the real stride.
Therefore the video snapshot for 720x480 are distorted since
 the buffer is 736x480

Solution: Make sure that if the buffer declares a stride we do
use that as the real width of the buffer

Change-Id: I8a05db4ff62a92fe4726e78a78ca6668a3a79fe3
Signed-off-by: Miguel Verdu <miguel.verdu@intel.com>
Reviewed-on: http://android.intel.com:8080/68229
Reviewed-by: Vehmanen, Kai <kai.vehmanen@intel.com>
Reviewed-by: Koski, Anttu <anttu.koski@intel.com>
Tested-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>