libcamera2: set parameters during initialization
BZ: 52165

Problem:
Due to CTS, default setting for flash is different than what it is in
the default scene mode (auto). Due to the inconsistency, setting the
flash off didn't propagate to 3A during the first setParameter call,
because libcamera2 thinks the flash setting has not changed.

Solution:
This patch sets all parameters during initialization and does not
check whether the setting changed or not during init(). The exception
is scene mode, which can't be set during init() due to the
inconsistency.

Change-Id: Id4fc1e38c71bc3adfc0a6cb406da1c56223aefea
Signed-off-by: Ukri Niemimuukko <ukri.niemimuukko@intel.com>
Reviewed-on: http://android.intel.com:8080/68692
Reviewed-by: Vtminen, LassiX <lassix.vtminen@intel.com>
Reviewed-by: Vehmanen, Kai <kai.vehmanen@intel.com>
Reviewed-by: Verdu, Miguel <miguel.verdu@intel.com>
Tested-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>