libcamera2: limit video preview with ov8830 sensor
BZ: 55640

With certain sensors (e.g. the OV8830 in CTP PR1), the ISP will run out
of viewfinder processing time at high video resolutions. In practise,
this shows as FPS not hitting the configured 30FPS.

To workaround this, extend the ISP limitations logic in AtomISP with
sensor specific limitations, and specifically for the "ov8830" sensor.

Change-Id: I524a6f110dd034aab8d4beb604d4a3c773291466
Signed-off-by: Kai Vehmanen <kai.vehmanen@intel.com>
Reviewed-on: http://android.intel.com:8080/69420
Reviewed-by: Tuominen, TeemuX <teemux.tuominen@intel.com>
Reviewed-by: Koski, Anttu <anttu.koski@intel.com>
Tested-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>