[PORT FROM JBMR0] libcamera2: handle postviewDone CB in case of VideoSnapshot
BZ: 64947

In Video Snapshot case the postview frame is actually a preview
frame. In this case the data is not in the usual location
but in AtomBuffer.GfxData field.
When processing the postview done callback in this case the
code was not ready to accept this types of frames, causing a
dereference of NULL.

Solution: create an empty camera_memory_t and set the data
and size fields to the ones from the Gfx buffer. In this way
we can present this Gfx buffer as a regular one
This is done only temporarily for the CB

Change-Id: Ib74cb10977fe7ef954044b4ddccefb12d6881ca5
Orig-Change-Id: I732d58d095e7eda2ba1d8437f72e43f388e463bc
Signed-off-by: Miguel Verdu <miguel.verdu@intel.com>
Reviewed-on: http://android.intel.com:8080/82885
Reviewed-by: Koski, Anttu <anttu.koski@intel.com>
Tested-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>