[PORT FROM MAIN]  libcamera2: Enzo: add scaling before encoding main image
BZ: 62545

For enzo front camera we run the sensor always in preview
In case the configured capture resolution is diferent than
the preview one we need to add a scaler before encoding.

This is done in a similar fashion as the thumbnail scaling

now the target resolutions for main and thumbnail are setup
during the initialize() call.
After that if the picture thread receives buffers to encode
that are of different resoution the scaling will take place

Change-Id: I650a42e0d2c524b09cea645e5eca776ea07a1da6
Orig-Change-Id: I83c66d4f8f26b6e66a3fbc53d97ae56238a7d0a9
Signed-off-by: Miguel Verdu <miguel.verdu@intel.com>
Signed-off-by: Lassi Väätämöinen <lassix.vtminen@intel.com>
Reviewed-on: http://android.intel.com:8080/74577
Reviewed-by: Koski, Anttu <anttu.koski@intel.com>
Tested-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>
Reviewed-on: http://android.intel.com:8080/76161
Tested-by: Tong, BoX <box.tong@intel.com>
Reviewed-by: Tong, BoX <box.tong@intel.com>