libcamera2: resetting values of smart scene on start
BZ: 61143

Issue:
When having one scene with hdrhint and smart scene
is turned off and then on, the same pattern is analyse and
no change are detected, so no callback are sent to the app
so hdrhint is not updated on the app side,

Solution:
This patch is resetting the values for hdr hint
and smart scene detected.
This allows to force to send a callback to the app
when the smart scene get restarted.

Change-Id: I1a6f84dbd4f62e1bd5b9a79d19e6d2f11de6a64f
Signed-off-by: Arnaud Mandy <arnaud.mandy@intel.com>
Reviewed-on: http://android.intel.com:8080/70419
Reviewed-by: Alanen, Juha <juha.alanen@intel.com>
Reviewed-by: Vehmanen, Kai <kai.vehmanen@intel.com>
Reviewed-by: Vtminen, LassiX <lassix.vtminen@intel.com>
Reviewed-by: Lampila, KalleX <kallex.lampila@intel.com>
Reviewed-by: Yang, Liang L <liang.l.yang@intel.com>
Reviewed-by: Verdu, Miguel <miguel.verdu@intel.com>
Reviewed-by: Koski, Anttu <anttu.koski@intel.com>
Tested-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>