libcamera2: handle command CAMERA_CMD_ENABLE_FOCUS_MOVE_MSG
BZ: 63679

Amendment to change Ib76d149a72cac92756ec27254f0985d95723a90d.
Previously the CAF callback was enabled, but due to the command
handler not being enabled we would end up always sending the callback
message to the framework, which would then intercept it. Thus, the
situation seems normal from applications' point of view. Instead,
we should be consistent with the framework state in HAL, and not send
the callback to FW, unless requested. This patch addresses the issue.

Change-Id: If32d6e7c15090c6f2c6052c9c95cefa3448563fc
Signed-off-by: Lassi Väätämöinen <lassix.vtminen@intel.com>
Reviewed-on: http://android.intel.com:8080/71055
Reviewed-by: Niemimuukko, Ukri <ukri.niemimuukko@intel.com>
Reviewed-by: Lampila, KalleX <kallex.lampila@intel.com>
Reviewed-by: Verdu, Miguel <miguel.verdu@intel.com>
Reviewed-by: Koski, Anttu <anttu.koski@intel.com>
Tested-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>