libcamera2: flush extra 3A processes when stream is lagging
BZ: 55065

In case AAAThread doesn't have time to process frames by time,
process requests start to pile up in the message queue. AAAThread
was never flushed but left working on the old frames over the
ControlThreads state changes.

By using direct proprietary ioctl, the processing context of 3A
and DVS will read the statistics of most recent frame. These do
not correspond with the processed timestamp when running late.
Running late also means that 3A processes are not done for each
frame.

This patch flushes the pile of newFrame-requests whenever frame
gets processed and replaces the timestamp with the most recent.
This should not occur normally, but with high cpu burden it
provides an new idicator for a performance problem, reduces extra
weight and gives faster recovery.

Change-Id: I2d07a27e10d2428c05da953062c2e86a43559252
Signed-off-by: Teemu Tuominen <teemux.tuominen@intel.com>
Reviewed-on: http://android.intel.com:8080/71882
Reviewed-by: Vehmanen, Kai <kai.vehmanen@intel.com>
Reviewed-by: Kantola, Lasse <lasse.kantola@intel.com>
Reviewed-by: Wada, Yoshio <yoshio.wada@intel.com>
Reviewed-by: Verdu, Miguel <miguel.verdu@intel.com>
Reviewed-by: Koski, Anttu <anttu.koski@intel.com>
Tested-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>