libcamera2: add PerformanceTracer for the ShutterLag
BZ: 52403

Adds log traces for ShutterLag KPI.

Measures the time between CameraHALs takePicture()-call and the
timestamp of picture taken.

One should note that actual shutterlag is typically measured with
realtime clock for the time between the shutter-press to the
image taken time. HAL's takePicture() may get delayed from the
shutter-press (e.g. for focusing) and the timestamp on other end
is written by the kernel after the image comes out from the ISP.

With knowing these facts, the log tracer can be used as a help
in identifying clear differences to the KPI.

Change-Id: I4b9aa60d57036fffdd7bb22c18dbd82730ad9b47
Signed-off-by: Teemu Tuominen <teemux.tuominen@intel.com>
Reviewed-on: http://android.intel.com:8080/71842
Reviewed-by: Yi, Hongyu <hongyu.yi@intel.com>
Reviewed-by: Yang, Liang L <liang.l.yang@intel.com>
Reviewed-by: Zheng, Hongjiang A <hongjiang.a.zheng@intel.com>
Reviewed-by: Verdu, Miguel <miguel.verdu@intel.com>
Reviewed-by: Niemimuukko, Ukri <ukri.niemimuukko@intel.com>
Reviewed-by: Vehmanen, Kai <kai.vehmanen@intel.com>
Reviewed-by: Koski, Anttu <anttu.koski@intel.com>
Tested-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>