libcamera2: hardcode max snapshot size in platform data
BZ: 64232

Recent patches in kernel have made return value of the IOTCL
VIDIOC_TRY_FMT not to reflect the maximum sensor size but
the maximum size the ISP can provide.

This information was used by the hal so now it needs to be
harcoded on a per platform basis.
We move this information to PlatformData

Change-Id: Ibeb2326c7e07ad7db4c017ffc28e78eab3290525
Signed-off-by: Miguel Verdu <miguel.verdu@intel.com>
Reviewed-on: http://android.intel.com:8080/71873
Reviewed-by: Thilak, Vimal <vimal.thilak@intel.com>
Reviewed-by: Koski, Anttu <anttu.koski@intel.com>
Tested-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>