libcamera2: reduce preview buffers count in still mode for L2P
BZ: 65846

Issue:
When the allocating preview buffer from GFX and qbuf
preview buffers to ISP, causes unnessary time consumption,
It's a point for optimization on L2P.

Solution:
The patch is for optimization on l2p.
Reduce the preview buffers count from 9 to 6 in still mode,
but in recording mode HAL still keeps 9 coupleBuffers, contains
9 preview buffers and 9 recording buffers, the design should
guarantee the performance for the live effects feature of
video component as well.

Change-Id: I6173568f1a17aa968ccffe3b0b280d6c587eab56
Signed-off-by: hongjian <hongjiang.a.zheng@intel.com>
Reviewed-on: http://android.intel.com:8080/72435
Reviewed-by: buildbot <buildbot@intel.com>
Reviewed-by: Verdu, Miguel <miguel.verdu@intel.com>
Reviewed-by: Zhu, ShaopingX <shaopingx.zhu@intel.com>
Reviewed-by: Vehmanen, Kai <kai.vehmanen@intel.com>
Tested-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: lab_cactus <lab_cactus@intel.com>
Tested-by: lab_cactus <lab_cactus@intel.com>