[PORT FROM MAIN]  libcamera2: video frame dropping based on setPreviewFrameRate()
BZ: 65853

MediaRecorder uses the deprecated setPreviewFrameRate() call for
setting the video FPS. This is not a problem in the normal application
case, when 30 FPS is assumed for each video quality, but for example in
Messaging attachments, when QCIF time lapse video is chosen to be recorded,
we need to provide the frames based on the MediaRecorder requrirement.

This patch fixes the issue by using video frame dropping, in case frame rate
is set.

Change-Id: I8f40936f9432228f7b06354d85c8960338bb0f6e
Signed-off-by: Lassi Väätämöinen <lassix.vtminen@intel.com>
Reviewed-on: http://android.intel.com:8080/73399
Reviewed-by: lab_cactus <lab_cactus@intel.com>
Reviewed-by: Lampila, KalleX <kallex.lampila@intel.com>
Reviewed-by: Verdu, Miguel <miguel.verdu@intel.com>
Reviewed-by: Hu, Gang A <gang.a.hu@intel.com>
Reviewed-by: Vehmanen, Kai <kai.vehmanen@intel.com>
Tested-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>
Reviewed-on: http://android.intel.com:8080/76147
Reviewed-by: cactus <cactus@intel.com>
Tested-by: Tong, BoX <box.tong@intel.com>
Reviewed-by: Tong, BoX <box.tong@intel.com>