[PORT FROM MAIN] libcamera2: generate burst length and fps setting by platform
BZ: 68845

Issue:
For R4, we don't have the code to get the burst
length and fps according to the platform's fps setting.
We should do the porting from R3 to R4.

Solution:
Do the porting for burst length and fps code from R3 to R4.
Actually, this is not one porting, this should be one new implementation.
We have re-designed the burst fps.
In the near future, we will use "Fast", "Medium" and "Low"
to instead of the real fps value. At that time, we need to
remove the fps caculation formula.

Depends on:
None.

Change-Id: If22fcbdcf9b4f6381762288139a9e0e996951a09
Signed-off-by: Yang Liang <liang.l.yang@intel.com>
Reviewed-on: http://android.intel.com:8080/74561
Reviewed-by: lab_cactus <lab_cactus@intel.com>
Reviewed-by: Verdu, Miguel <miguel.verdu@intel.com>
Reviewed-by: Mandy, Arnaud <arnaud.mandy@intel.com>
Tested-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>
Reviewed-on: http://android.intel.com:8080/77461
Reviewed-by: Hu, Gang A <gang.a.hu@intel.com>
Tested-by: Hu, Gang A <gang.a.hu@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>