[PORT FROM MAIN] libcamera2: get the ISP device's name by according to the platform
BZ: 68828

Issue:
The original issue is that the camera will crash when we take picture with focus bracket enabled.
The root cause is that the Lex's device name is not the same as the PR3's.
The device name is platform related.

Solution:
Move the ISP device's name from the AtomISP.cpp to
the PlatformMedfield.h and the PlatformClovertrail.h.
Then we can get the ISP device's name by the method getDeviceName() in the PlatformData.cpp

Depends on:
None.

Change-Id: Iec488bb6964751da4c3e923ddd56b2b2ceba8525
Signed-off-by: Yang Liang <liang.l.yang@intel.com>
Reviewed-on: http://android.intel.com:8080/74802
Reviewed-by: lab_cactus <lab_cactus@intel.com>
Reviewed-by: Gong, Shuguang <shuguang.gong@intel.com>
Reviewed-by: Verdu, Miguel <miguel.verdu@intel.com>
Reviewed-by: Liu, Zhengao <zhengao.liu@intel.com>
Reviewed-by: Hu, Gang A <gang.a.hu@intel.com>
Reviewed-by: Vehmanen, Kai <kai.vehmanen@intel.com>
Tested-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>
Reviewed-on: http://android.intel.com:8080/77462
Tested-by: Hu, Gang A <gang.a.hu@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>