libcamera2: implement render preview via HW overlay
BZ: 41619

In order to introduce support for preview rendering via
overlay we introduce a new level of abstraction in the
PreviewThread.
The new class PreviewMessageHandler is a base class
where we have factored out the common parts of the PreviewThread
that were used when rendering via Gfx plane or Overlay.
Then we have 2 new derived classes where we implement the specific
buffer management for each method.

There are also fixes in AtomISP to fix the following issues:
- correct initialization of preview buffers (format was missing)
- usage of correct number of preview buffers in video mode. It
  has to be the same number as encoding buffers.

By default overlay rendering is not enabled. Later patches
will enable the overlay rendering path

Change-Id: I0367c93f3a8b08ad4ee8ef2752ae8459c144b422
Signed-off-by: Miguel Verdu <miguel.verdu@intel.com>
Reviewed-on: http://android.intel.com:8080/78243
Reviewed-by: Koski, Anttu <anttu.koski@intel.com>
Tested-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>