[PORT FROM MAIN] libcamera2: add support for continuous capture to PlatformData
BZ: 72233

Continuous capture mode is sensitive to SoC memory subsystem performance
and configuration, and it cannot be enabled with all hardware
combinations.

Add a SPID based check to CTP RHB platform data so that continuous
capture is only enabled for RHB PR1.9 and newer. Continuous capture is
disabled for all other platforms for now.

Change-Id: Icc3c495c5a08e7dc2168d16bafa866ff5242976d
Signed-off-by: Kai Vehmanen <kai.vehmanen@intel.com>
Reviewed-on: http://android.intel.com:8080/80954
Reviewed-by: Koski, Anttu <anttu.koski@intel.com>
Tested-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>