[PORT FROM R4.1] libcamera2: The EXIF of Digital Zoom is configured to be confused and wrong
BZ: 70221

Issue:
The EXIF of digital zoom is showing with wrong opened by some image
professional softwhare, such as XnView. For example, the normal is
showed with 0.1, and the max zoom is showed with 1.6.

Solution:
Since the denominator is set as 10, the numerator is not set to divide 10.
Otherwise, the result would reduce 10 times - that is not expected to result
the wrong zoom value.

Change-Id: If5d9d02c5c4246112e0eb63965e07a61f348d562
Orig-Change-Id: I4f05850b419da646da284966d8ea807a0c84b72e
Signed-off-by: Liu Zhengao <zhengao.liu@intel.com>
Reviewed-on: http://android.intel.com:8080/85016
Reviewed-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: Yang, Liang L <liang.l.yang@intel.com>
Reviewed-by: Verdu, Miguel <miguel.verdu@intel.com>
Tested-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>