[PORT FROM MAIN] libcamera2: fix switch from continuous preview to HDR mode
BZ: 70521

Fix the case when HDR is enabled when preview is already running.
This failed as setting HDR related parameters did not trigger
a mode change, and HDR was attempted in continuous mode.

This patch also removes preallocation of the HDR context, as the
allocations do not persist over a ISP restart. This happens
often when continuous mode is available as an option.

Change-Id: Ia882279eb23fa8a1b6eca966f347831be9fecc3b
Signed-off-by: Kai Vehmanen <kai.vehmanen@intel.com>
Reviewed-on: http://android.intel.com:8080/80959
Reviewed-by: Koski, Anttu <anttu.koski@intel.com>
Tested-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>