[PORT FROM MAIN] libcamera2: HDR: JpegHwEncoder: clear surface before restore
BZ: 69915

The new surface created for out buffer is not deleted.
Which leads in having 3 different surfaces (input) + 1 out
which gives 4 surfaces.
This will prevent the encoder to reset the context next time
since all surfaces have a buffer.
This can create error when encoding the composed image.

In this patch, we clear the buffer to surface vector before
restoring the original setting to make sure the HDR will be
working on the next HDRcapture.

Change-Id: Ic093933fcf838c263c8a9f84176ffde28084003f
Signed-off-by: Arnaud Mandy <arnaud.mandy@intel.com>
Reviewed-on: http://android.intel.com:8080/81651
Reviewed-by: Koski, Anttu <anttu.koski@intel.com>
Tested-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>