[PORT FROM MAIN] libcamera2: have a guard period with ZSL snapshot offset
BZ: 73055

Application is able to place takePicture() request in queue
right after startPreview(). When occurs, CSS1.5 ZSL implementation
resulted in picking the very first frame from the sensor where
settings are not applied.

Logically continous capture would need the 1st-frame skipping logic
when streaming is started, as is in use with mode-switch architecture.

This change initializes continuous capture with 0-offset and
updates the value to -1 with preview frames done-phase. Hence, with Gfx
preview buffers in use, offset is kept 0 for the first 3 preview frames.
Snapshots taken with 0-offset effectively removes the occurences of the
bug. Certainty further depends on CSS.

Change-Id: I9b07815fdb04704db5279344c3e05572ac614f3e
Signed-off-by: Teemu Tuominen <teemux.tuominen@intel.com>
Reviewed-on: http://android.intel.com:8080/80960
Reviewed-by: Koski, Anttu <anttu.koski@intel.com>
Tested-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>