[PORT FROM JBMR0] libcamera2: separate panorama stitching to its own thread
BZ: 73469

Problem: new panorama library runs the stitching in the stitch
function caller context. Old library used to return immediately.
As a result, ControlThread is blocked for a long time, as is
PanoramaThread, which will be unable to continue processing
overlap of preview images or to trigger new captures.

Solution: this patch adds a new thread which runs the stitching,
so that preview can be started immediately as it used to. An
unfortunate consequence is the need to make a copy of the capture.
A best effort memory allocation is done for the copy and the
stitches are queued. With the new thread, ControlThread and
PanoramaThread are unblocked for the period of stitching, and
multiple captures can be queued up for stitching as previously with
the old panorama library.

Change-Id: I9a57b83b6db9c18f39231df07302d38a011a8f30
Orig-Change-Id: If4f38dbbfd9cda77619f757faeba4b8d5c5c9233
Signed-off-by: Ukri Niemimuukko <ukri.niemimuukko@intel.com>
Reviewed-on: http://android.intel.com:8080/82875
Reviewed-by: Koski, Anttu <anttu.koski@intel.com>
Tested-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>