[PORT FROM JBMR0] libcamera2: AtomISP: reset driver state when starting preview
BZ: 73066

Switching between continuous, non-continuous and acceleration
modes is not reliable as the driver state is not reset to initial
settings between different modes. BZ 74007 is used to track
the problem on CSS side.

Previously we did the reset after stopPreview, but this breaks
use-cases that expect camera hardware to maintain its state
when moving from non-continuous preview to capture. For example
focus is not maintained from preview to capture.

This change is limited to platforms that support continuous
capture mode, as older CSS firmwares do not trigger this problem.

Change-Id: Iffe4cd18f5e4ea57ba5160af70c0cdd04e62384b
Orig-Change-Id: I2c7d4c2845412d79c35f6b78fea3afa82091098a
Signed-off-by: Kai Vehmanen <kai.vehmanen@intel.com>
Reviewed-on: http://android.intel.com:8080/82880
Reviewed-by: Koski, Anttu <anttu.koski@intel.com>
Tested-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>