libcamera2: [REVERTME] 1080P to 1024x576 image scaler for panorama
BZ: 74960

Problem: When ZSL is enabled, 1024x576 postview images are corrupted.

This patch adds a 1080P to 1024x576 image scaler, and it is used when
ZSL is enabled with Panorama to recreate the postview image after
snapshot capture.

This patch should be reverted after the ZSL postview image corruption
is sorted out properly.

Change-Id: I57d83c74a7cd4c15466e99e64188be2f5e4f4856
Signed-off-by: Ukri Niemimuukko <ukri.niemimuukko@intel.com>
Signed-off-by: Timo Kaikumaa <timo.kaikumaa@intel.com>
Reviewed-on: http://android.intel.com:8080/81676
Reviewed-by: Koski, Anttu <anttu.koski@intel.com>
Tested-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>