libcamera2: let cpf control assist light behaviour
BZ: 73761

This patch makes CPF parameter to be in control of the assist light
behaviour, as it should. In practice the categorical turning on
of the assist light for CAM_AE_FLASH_MODE_ON is removed, and AAA
is queried for the assist light need. AAA decision then depends
on the CPF parameter.

Change-Id: Ib4845f1d77e21a218463b6144fde5261e510a604
Signed-off-by: Ukri Niemimuukko <ukri.niemimuukko@intel.com>
Reviewed-on: http://android.intel.com:8080/82022
Reviewed-by: Krestyannikov, Evgeny <evgeny.krestyannikov@intel.com>
Reviewed-by: Vehmanen, Kai <kai.vehmanen@intel.com>
Reviewed-by: Koski, Anttu <anttu.koski@intel.com>
Tested-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>