libcamera2: TMP HACK: Disable overlay for front camera
BZ: 41619

Temporarily disable overlay rendering for front camera.
This is done by making the overlay rendering a per camera
feature. This is not needed in general.
But it is needed now until we have the optimized 90
degree routine for front camera. This current routine
for the back camera is anticlock-wise and for the
front camera we also need the clockwise

Once we have that routine the mPreviewViaOverlay
will go back to be a platform feature.

Change-Id: I00847cc1c80e598578472bf7cc1f8e51b1ebec72
Signed-off-by: Miguel Verdu <miguel.verdu@intel.com>
Reviewed-on: http://android.intel.com:8080/82131
Reviewed-by: Koski, Anttu <anttu.koski@intel.com>
Tested-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>