[PORT FROM R4-ENZO] libcamera2: using public AF mode for CAF callback status
BZ: 76004

Previously we used the HAL internal AF status when deciding when to
send the Continuous AF callback to the app. We should honour the
application-set "public" AF mode instead, which makes sense from
apps' point-of-view, whereas in HAL we might use different AF mode
internally, like "touch" or "face" modes).

This patch addresses the issue.

Change-Id: I7ad2eb3d4f6b36f248555910f04923480c3856df
Orig-Change-Id: I43f7a20bcbac311d690a7cab3642a988a0001be9
Signed-off-by: Lassi Väätämöinen <lassix.vtminen@intel.com>
Reviewed-on: http://android.intel.com:8080/83257
Reviewed-by: Koski, Anttu <anttu.koski@intel.com>
Tested-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>