libcamera2: fixes face indicator flash after touch focus
BZ: 69205

The face indicator was briefly shown on UI due to extra face detection
callback sent when starting face detection after coming out
of touch focus, and faces being detected before doing touch focus.
This was due to us not sending frames to FD library
during touch focus, hence it was in invalid state. This patch
implements reset for the FD library that fixes the issue.

Change-Id: I7e27d9c7dbaf3d79e667ee96a3631ef7ed6cd913
Signed-off-by: Lassi Väätämöinen <lassix.vtminen@intel.com>
Reviewed-on: http://android.intel.com:8080/84730
Reviewed-by: Verdu, Miguel <miguel.verdu@intel.com>
Reviewed-by: Vehmanen, Kai <kai.vehmanen@intel.com>
Reviewed-by: Koski, Anttu <anttu.koski@intel.com>
Tested-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>