libcamera2: ensure that each detected face has unique ID
BZ: 65361

Face ID should be unique for each detected face, so fill the face ID with
tracking ID if FR is not enabled or when the face is not recognized.
Tracking ID is always unique for each detected face. Use negative values
so that the application doesn't misinterpret the face as recognized.

Change-Id: I4761c687a0454a5e1dc93a7e10bef9d9727c1a2e
Signed-off-by: Juha Alanen <juha.alanen@intel.com>
Reviewed-on: http://android.intel.com:8080/84895
Reviewed-by: Heo, Chungbin <chungbin.heo@intel.com>
Reviewed-by: Verdu, Miguel <miguel.verdu@intel.com>
Tested-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>