[PORT FROM R4.1] libcamera2: set the recording buffer number by platform
BZ: 66003

Issue:
To enable the video feature: video effects feature, it needs more
video buffers in the camera HAL.
The video effects is not one POR for yukkabeach and lexington.
If we can disable this feature, we can reduce the video buffers number
in the camera HAL.
In one previous patch, all recording buffer number has been set to 9.
This is not right.

Solution:
set the recording buffer number by platform.
Set it to 6 if the platform is yukkabeach and lexington.
Set it to 9 if the platform is MFLD and CTP.

Depends on:
85519

Change-Id: I39530a8cec2c0b6a318254af1c335641f8ac5b93
Orig-Change-Id: Ieea179df9d323fb74bb2676f435123c85c6c9dfe
Signed-off-by: Yang Liang <liang.l.yang@intel.com>
Reviewed-on: http://android.intel.com:8080/85522
Reviewed-by: Koski, Anttu <anttu.koski@intel.com>
Tested-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: Verdu, Miguel <miguel.verdu@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>