libcamera2: host preview state from PreviewThread
BZ: 73975

This is refactoring and additional work continued from change
http://android.intel.com:8080/84944
"libcamera2: introduce active stopped state for ZSL shot2shot"

Moves preview sub-states to the PreviewThread. This is to be
able to provide generality over newly added state transitions in
CSS1.5 continuous mode. Meaningly to control preview behaviour
more generally over capturing sequences that differ in
applications.

Distinct booleans for "active stopped" state are removed and
preview is generally kept hidden after capture. When possible
(internal preview buffers in use), continuous-mode is allowed
to stay running when preview is explicitly stopped.

Current PreviewThread is acting as a streaming buffer provider and
mediator, where it originally did not need a state. As pre-planned,
PreviewThread will be converted into component responsible of the
whole preview stream and utilize further its states in internal
operations.

Effectively the sequences with continuous takePicture() are:

IntelCamera2 ZSL:
- takePicture() -> preview stopped (hidden)
- setPreviewWindow() -> window handle doesn't change (supported)
- startPreview() -> preview enabled only by changing a flag

Google Camera ZSL:
- takePicture() -> preview stopped (hidden)
- stopPreview() -> PreviewThread stopped, buffers released
- setPreviewWindow() -> window handle changes (supported)
- startPreview() -> PreviewThread re-configured

To distinct modes in startPreview() bases on whether the external
buffers are configured in AtomISP. The fuction for this purpose
is added.

Change-Id: I834ef261a629cc6aba649f315176d70d94067fec
Signed-off-by: Teemu Tuominen <teemux.tuominen@intel.com>
Reviewed-on: http://android.intel.com:8080/87093
Reviewed-by: Koski, Anttu <anttu.koski@intel.com>
Tested-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>